+++ locallib.php
@@ -42,6 +42,7 @@ define('ATT_SORT_FIRSTNAME', 2);
 define('ATTENDANCE_AUTOMARK_DISABLED', 0);
 define('ATTENDANCE_AUTOMARK_ALL', 1);
 define('ATTENDANCE_AUTOMARK_CLOSE', 2);
+
 /**
  * Get statuses,
  *
@@ -709,4 +710,114 @@ SELECT a.id, a.course as courseid, c.fullname as coursename, atl.studentid AS us
                 ) p GROUP by courseid, coursename {$orderby}";
 
     return $DB->get_records_sql($sql, $params);
+}
+
+/**
+ * Generates a list of users flagged at-risk.
+ *
+ * @param array $courseids optional list of courses to return
+ * @param array $sincetime optional allows a list to be calculated for cron processing.
+ * @param bool $allfornotify get notification list for scheduled task.
+ * @return stdClass
+ */
+function attendance_get_users_to_notify($courseids = array(), $orderby = '', $sincetime = 0, $allfornotify = false) {
+    global $DB;
+
+    $joingroup = 'LEFT JOIN {groups_members} gm ON (gm.userid = atl.studentid AND gm.groupid = ats.groupid)';
+    $where = ' AND (ats.groupid = 0 or gm.id is NOT NULL)';
+    $params = array();
+
+    if (!empty($courseids)) {
+        list($insql, $inparams) = $DB->get_in_or_equal($courseids, SQL_PARAMS_NAMED);
+        $where .= ' AND c.id ' . $insql;
+        $params = array_merge($params, $inparams);
+    }
+    if ($allfornotify) {
+        // Exclude warnings that have already been sent.
+        $where .= ' AND ns.id IS NULL ';
+    }
+
+    $unames = get_all_user_name_fields(true);
+    $unames2 = get_all_user_name_fields(true, 'u');
+
+    $idfield = $DB->sql_concat('cm.id', 'atl.studentid', 'n.id');
+    $sql = "SELECT {$idfield} as uniqueid, a.id as aid, {$unames2}, a.name as aname, cm.id as cmid, c.id as courseid,
+                    c.fullname as coursename, atl.studentid AS userid, n.id as notifyid, n.warningpercent, n.emailsubject,
+                    n.emailcontent, n.emailcontentformat, n.emailuser, n.thirdpartyemails, ns.timesent, n.warnafter,
+                     COUNT(DISTINCT ats.id) AS numtakensessions, SUM(stg.grade) AS points, SUM(stm.maxgrade) AS maxpoints,
+                      SUM(stg.grade) / SUM(stm.maxgrade) AS percent
+                   FROM {attendance_sessions} ats
+                   JOIN {attendance} a ON a.id = ats.attendanceid
+                   JOIN {course_modules} cm ON cm.instance = a.id
+                   JOIN {course} c on c.id = cm.course
+                   JOIN {modules} md ON md.id = cm.module AND md.name = 'attendance'
+                   JOIN {attendance_log} atl ON (atl.sessionid = ats.id)
+                   JOIN {user} u ON (u.id = atl.studentid)
+                   JOIN {attendance_statuses} stg ON (stg.id = atl.statusid AND stg.deleted = 0 AND stg.visible = 1)
+                   JOIN {attendance_warning} n ON n.idnumber = cm.id
+                   LEFT JOIN {attendance_warning_done} ns ON ns.notifyid = n.id AND ns.userid = atl.studentid
+                   JOIN (SELECT attendanceid, setnumber, MAX(grade) AS maxgrade
+                           FROM {attendance_statuses}
+                          WHERE deleted = 0
+                            AND visible = 1
+                         GROUP BY attendanceid, setnumber) stm
+                     ON (stm.setnumber = ats.statusset AND stm.attendanceid = ats.attendanceid)
+                  {$joingroup}
+                  WHERE ats.sessdate >= {$sincetime} {$where}
+                    AND ats.lasttaken != 0
+                GROUP BY uniqueid, a.id, a.name, a.course, c.fullname, atl.studentid, n.id, n.warningpercent,
+                         n.emailsubject, n.emailcontent, n.emailcontentformat, n.warnafter,
+                         n.emailuser, n.thirdpartyemails, ns.timesent, cm.id, c.id, {$unames2}
+                HAVING n.warnafter <= COUNT(DISTINCT ats.id) AND n.warningpercent > ((SUM(stg.grade) / SUM(stm.maxgrade)) * 100)
+                      {$orderby}";
+
+    if (!$allfornotify) {
+        $idfield = $DB->sql_concat('cmid', 'userid');
+        // Only show one record per attendance for teacher reports.
+        $sql = "SELECT {$idfield} as id, {$unames}, aid, cmid, courseid, aname, coursename, userid, MIN(warningpercent),
+                        numtakensessions, points, maxpoints, percent, timesent
+              FROM ({$sql}) as m
+         GROUP BY id, aid, cmid, courseid, aname, userid, numtakensessions, points, maxpoints,
+                  percent, coursename, timesent, {$unames} {$orderby}";
+    }
+
+    return $DB->get_records_sql($sql, $params);
+
+}
+
+/**
+ * Template variables into place in supplied email content.
+ *
+ * @param object $record db record of details
+ * @return array - the content of the fields after templating.
+ */
+function attendance_template_variables($record) {
+    $templatevars = array(
+        '/%coursename%/' => $record->coursename,
+        '/%courseid%/' => $record->courseid,
+        '/%userfirstname%/' => $record->firstname,
+        '/%userlastname%/' => $record->lastname,
+        '/%userid%/' => $record->userid,
+        '/%warningpercent%/' => $record->warningpercent,
+        '/%attendancename%/' => $record->aname,
+        '/%cmid%/' => $record->cmid,
+        '/%numtakensessions%/' => $record->numtakensessions,
+        '/%points%/' => $record->points,
+        '/%maxpoints%/' => $record->maxpoints,
+        '/%precent%/' => $record->percent,
+    );
+    $extrauserfields = get_all_user_name_fields();
+    foreach ($extrauserfields as $extra) {
+        $templatevars['/%'.$extra.'%/'] = $record->$extra;
+    }
+    $patterns = array_keys($templatevars); // The placeholders which are to be replaced.
+    $replacements = array_values($templatevars); // The values which are to be templated in for the placeholders.
+    // Array to describe which fields in reengagement object should have a template replacement.
+    $replacementfields = array('emailsubject', 'emailcontent');
+
+    // Replace %variable% with relevant value everywhere it occurs in reengagement->field.
+    foreach ($replacementfields as $field) {
+        $record->$field = preg_replace($patterns, $replacements, $record->$field);
+    }
+    return $record;
 }