+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### Date:		2017-June-14
 ### Release:    3.3.8
 - New Feature: Allow automatic marking using site logs.
+- New Feature: Warn users when attendance drops below threshold.
 - Improvement: Allow default view for teachers to be set at admin level.
 - Improvement: All courses user report now displays as table.
 - Bug fix: Restored attendances do not create calendar events correctly.